move foundational models to core/
