--- backend/authentication/models.py
@@ -1,39 +1,0 @@
-from django.contrib.auth.models import AbstractUser
-from django.db import models
-from django.contrib.auth.base_user import BaseUserManager
-
-class CustomUserManager(BaseUserManager):
-    def create_user(self, email, password=None, **extra_fields):
-        if not email:
-            raise ValueError('The Email must be set')
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password=None, **extra_fields):
-        extra_fields.setdefault('is_staff', True)
-        extra_fields.setdefault('is_superuser', True)
-
-        return self.create_user(email, password, **extra_fields)
-
-class CustomUser(AbstractUser):
-    is_verified = models.BooleanField(default=False)
-    email = models.EmailField(unique=True)
-    verification_token = models.CharField(max_length=100, blank=True, null=True)
-    organization = models.ForeignKey(
-        'access_control.Organization',
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name='users'
-    )
-
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = ['username']
-
-    objects = CustomUserManager()
-
-    def __str__(self):
-        return self.email
